Use data container to send data
