--- Assets/Scripts/DataCollection/JsonHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1ba7eda128ebee044bbe8a8ffc347086
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
